better yet

+++ app/src/main/java/com/example/jakeshepherd/tamago/MainActivity.java
@@ -3,6 +3,7 @@ package com.example.jakeshepherd.tamago;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
+import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
@@ -43,9 +44,11 @@ public class MainActivity extends AppCompatActivity{
         del.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
+                if (db.getNumberOfRows() == 0){
+                    Snackbar.make(view, "Nothing to delete", Snackbar.LENGTH_LONG).setAction("Action", null).show();
+                    return;
+                }
                 db.deleteRowData(db.getNumberOfRows());
-                Log.d("debug", "Number of Rows: " + db.getNumberOfRows());
-                Log.d("debug", "Food IDs: " + db.getAllFoodIDs());
                 fridge.removeAllViews();
                 updateFridge();
                 if(tester.getVisibility() == View.GONE && db.getNumberOfRows() == 0)
